% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{clean}
\alias{clean}
\title{Methods, grouped by type of request calling «clean».}
\usage{
clean(method, query, tidy = TRUE, ...)
}
\arguments{
\item{method}{Name of supported method. Use \code{supported_cleans()} to see all supported methods.}

\item{query}{Main request. See examples or \href{https://dadata.ru/api/}{https://dadata.ru/api/} for more.}

\item{tidy}{If TRUE returns response from the server as tidy \code{tibble}. Else, returns JSON as nested list.}

\item{...}{Specify additional calling parameters, if necessary. For more information on available parameters of
corresponding method, see the description page on \href{https://dadata.ru/api/}{https://dadata.ru/api/}}
}
\description{
Calls DADATA's REST API, with one of the available methods. All supported
methods related with «suggestion» you can see in \code{supported_cleans()}
}
\section{Warning}{

All methods presented in this section are for a fee. Learn more about costs on \href{https://dadata.ru/api/}{DADATA API Documentation}
}

\examples{
#save your tokens first
\dontrun{
save_dadata_tokens(api_token = '__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token = '__INSERT_YOUR_SECRET_TOKEN_HERE__')
}
# Validate and cleanse address
clean(method = "address", "мск сухонска 11/-89")
# Validate and cleanse name
clean(method = "name", query = "Срегей владимерович иванов")
# Validate and cleanse phone
clean(method = "phone", "раб 846)231.60.14 *139")
}
\seealso{
For more information about DADATA API visit the following link:
\href{https://dadata.ru/api/}{DADATA API Documentation}.
}
