% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{find_by_id}
\alias{find_by_id}
\title{Methods, grouped by type of request calling «findById».}
\usage{
find_by_id(method, query = NULL, tidy = TRUE, ...)
}
\arguments{
\item{method}{Name of supported method. Use \code{supported_find_by_id()} to see all supported methods.}

\item{query}{Main request. See examples or \href{https://dadata.ru/api/}{https://dadata.ru/api/} for more.}

\item{tidy}{If TRUE returns response from the server as tidy \code{tibble}. Else, returns JSON as nested list.}

\item{...}{Specify additional calling parameters, if necessary. For more information on available parameters of
corresponding method, see the description page on \href{https://dadata.ru/api/}{https://dadata.ru/api/}}
}
\description{
Calls DADATA's REST API, with one of the available methods. All supported
methods related with «suggestion» you can see in \code{supported_find_by_id()}
}
\examples{
#save your tokens first
\dontrun{
save_dadata_tokens(api_token = '__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token = '__INSERT_YOUR_SECRET_TOKEN_HERE__')
}
#Find address by FIAS ID
find_by_id(method = "address", query = "9120b43f-2fae-4838-a144-85e43c2bfb29")

#Find company by INN and KPP:
find_by_id(method = "party", query = "7707083893", kpp = "540602001")

#Tax office
find_by_id(method = "fns_unit", query = "5257")
}
\seealso{
For more information about DADATA API visit the following link:
\href{https://dadata.ru/api/}{DADATA API Documentation}.
}
