% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{get_dadata_tokens}
\alias{get_dadata_tokens}
\title{Retrieves the previously saved API tokens}
\usage{
get_dadata_tokens()
}
\value{
Returns a \code{data.frame} with the saved API tokens
}
\description{
Retrieves a \code{data.frame} with the available tokens
previously saved into the environment under the RDADATA variable
through the \code{save_dadata_tokens()} function.
}
\examples{
\dontrun{
save_dadata_tokens(api_token='__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token='__INSERT_YOUR_SECRET_TOKEN_HERE__')
get_dadata_tokens()
}
}
