% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{locate}
\alias{locate}
\title{Methods, grouped by type of request calling «locate» --- «geo» and «ip»}
\usage{
locate(method, tidy = TRUE, ...)
}
\arguments{
\item{method}{Name of supported method. Only «ip» or «geo»}

\item{tidy}{If TRUE returns response from the server as tidy \code{tibble}. Else, returns JSON as nested list.}

\item{...}{Specify additional calling parameters. For more information on available parameters of
corresponding method, see the description page on \href{https://dadata.ru/api/}{https://dadata.ru/api/}}
}
\description{
Calls DADATA's REST API, with one of the available methods.
}
\examples{
#save your tokens first
\dontrun{
save_dadata_tokens(api_token = '__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token = '__INSERT_YOUR_SECRET_TOKEN_HERE__')
}
# Identifies the city by its IP address in Russia. Uses the client's IP address
locate(method = "ip", ip = "46.226.227.20")
# request with a radius limit of 50 m
locate(method = "geo", lat = 55.601983, lon = 37.359486, radius_meters = 50)
}
\seealso{
For more information about DADATA API visit the following link:
\href{https://dadata.ru/api/}{DADATA API Documentation}.
}
