% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{personal_info}
\alias{personal_info}
\title{Methods, that shown your personal info}
\usage{
personal_info(method)
}
\arguments{
\item{method}{Name of supported method. Available 3 methods, without adjusted parameters:
\itemize{
\item stat
\item balance
\item version
}}
}
\description{
Calls available methods to check your personal info on DADATA
}
\examples{
\dontrun{
# show your stats of used methods
personal_info("stat")
# show your current balance
personal_info("balance")
# show actual time of available resources
personal_info("version")
}
}
