% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{save_dadata_tokens}
\alias{save_dadata_tokens}
\title{Saves credentials tokens in the environment}
\usage{
save_dadata_tokens(api_token = NULL, secret_token = NULL)
}
\arguments{
\item{api_token}{Token required for REST API authorization.
You can get the token on your personal profile page,
after logging in at \href{https://dadata.ru/profile/#info}{https://dadata.ru/profile/#info}.
It will be saved in the environment as API_TOKEN.}

\item{secret_token}{Token required for paid plans.
It will be saved in the environment as SECRET_TOKEN.}
}
\value{
Saves the tokens in the users environment - it does not return any object.
}
\description{
Saves credentials tokens
the users' environment. It has the advantage that it is not necessary
to explicitly publish the credentials in the users code. Just do it one
time and you are set. To update any of the parameters just save again and
it will overwrite the older credential.
}
\examples{
\dontrun{
save_dadata_tokens(api_token='__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token='__INSERT_YOUR_SECRET_TOKEN_HERE__')
}
}
