% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{suggest}
\alias{suggest}
\title{Methods, grouped by type of request calling «suggest».}
\usage{
suggest(method, query = NULL, tidy = TRUE, ...)
}
\arguments{
\item{method}{Name of supported method. Use \code{supported_suggestion()} to see all supported methods.}

\item{query}{Main request. See examples or \href{https://dadata.ru/api/}{https://dadata.ru/api/} for more.}

\item{tidy}{If TRUE returns response from the server as tidy \code{tibble}. Else, returns JSON as nested list.}

\item{...}{Specify additional calling parameters, if necessary. For more information on available parameters of
corresponding method, see the description page on \href{https://dadata.ru/api/}{https://dadata.ru/api/}}
}
\description{
Calls DADATA's REST API, with one of the available methods. All supported
methods related with «suggestion» you can see in \code{supported_suggestion()}
}
\examples{
#save your tokens first
\dontrun{
save_dadata_tokens(api_token='__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token='__INSERT_YOUR_SECRET_TOKEN_HERE__')
}
#Suggest country
suggest(method = "country", query = "та")

#Suggest company with spicified parametrs
suggest(method = "party", query = "сбер", branch_type = "MAIN", count = 20)

#Suggest bank
suggest(method = "bank", query = "ти")
}
\seealso{
For more information about DADATA API visit the following link:
\href{https://dadata.ru/api/}{DADATA API Documentation}.
}
