% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{supported_cleans}
\alias{supported_cleans}
\title{Returns the supported API names for cleans methods}
\usage{
supported_cleans()
}
\value{
Returns a character vector of all supported API names.
}
\description{
Function shows all  supported API names that are
available to choose for the \code{clean()} function.
}
