% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{supported_find_by_id}
\alias{supported_find_by_id}
\title{Returns the supported API names for find by ID methods}
\usage{
supported_find_by_id()
}
\value{
Returns a character vector of all the supported API names.
}
\description{
Function shows all the supported API names that are
available to choose for the \code{find_by_id()} function.
}
