% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_functions.R
\name{supported_suggestions}
\alias{supported_suggestions}
\title{Returns the supported API names for suggestions methods}
\usage{
supported_suggestions()
}
\value{
Returns a character vector of all the supported API names.
}
\description{
Function shows all the supported API names that are
available to choose for the \code{suggest()} function.
}
